Updates to run with the dynamic ip
As the phone changes the subnet of its tethering IP range on every boot,
I can not get away with the static IPs. Hence using dynamic IPs here.
Note that it requires HASS to have a working SSDP setup. In particular
this means configuring the multicast interface to be wlan1, which
happens to be the tethering interface used on lineageos.

Further I changed the logger to output to GPIO2 which is UART1 on
esp8285. It also happens to be the same pin as the internal LED, so I
can't use that for status_led any more. I think having a debug console
log is more useful (it also blinks the led whenever there is some
debug messages printed)